\name{go2bigq-package}
\alias{go2bigq-package}
\docType{package}
\title{ \packageTitle{go2bigq}
}
\description{
\packageDescription{go2bigq} 
}
\details{
\tabular{ll}{
Package: \tab go2bigq\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2022-06-02\cr
License: \tab GPL-3 \cr
}

}
\author{
Carl Witthoft

Maintainer:Carl Witthoft carl@witthoft.com

}

